Added the first few validity tests
